% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_200.R
\name{check_for_200}
\alias{check_for_200}
\title{Check for error messages on non-200 responses}
\usage{
check_for_200(response)
}
\arguments{
\item{response}{A response from a POST call to one of the oliver Sign-in Domains.}
}
\description{
Check for error messages on non-200 responses
}
