% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_poct.R
\name{extract_poct}
\alias{extract_poct}
\title{Extract token from Sprout sign-in GET response}
\usage{
extract_poct(response)
}
\arguments{
\item{response}{A response from a GET to one of the oliver Sign-in Domains containing a "poc.t" cookie.}
}
\description{
Extract token from Sprout sign-in GET response
}
