% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_info.R
\name{get_user_info}
\alias{get_user_info}
\title{Get authenticated user information from the Sprout API}
\usage{
get_user_info(user_data)
}
\arguments{
\item{user_data}{A `tbl``, most likely returned from
`post_auth_to_sprout()`, and then processed with `parse_json_content()`.}
}
\description{
This is essentially a configuration file for the columns needing
   needing to be selected from the user session object.
}
