% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json_content.R
\name{parse_json_content}
\alias{parse_json_content}
\title{Parse JSON content obtained from the Sprout API}
\usage{
parse_json_content(session_obj)
}
\arguments{
\item{session_obj}{A session object, most likely a list returned from
`post_auth_to_sprout()`.}
}
\description{
Parse JSON content obtained from the Sprout API
}
