% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_auth_to_sprout.R
\name{post_auth_to_sprout}
\alias{post_auth_to_sprout}
\title{Authenticate to the Sprout API}
\usage{
post_auth_to_sprout(email = NULL, password = NULL, env = NULL,
  user_dat = FALSE)
}
\arguments{
\item{email}{An email registered with the Sprout application.}

\item{password}{A password associated with `email`.}

\item{env}{The desired API environment.}
}
\description{
Authenticate to the Sprout API
}
