% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_subdomain.R
\name{return_subdomain}
\alias{return_subdomain}
\title{Define an Sprout subdomain to serve as a base url}
\usage{
return_subdomain(env = NULL)
}
\arguments{
\item{env}{A character string representing the desired API environment: 'production', 'usertest', or 'staging'.}
}
\description{
Define an Sprout subdomain to serve as a base url
}
